CommunityRequestsHooks: use static context in ::getCanonicalEntityPage
Use WishlistConfig::getCanonicalEntityPageRef() so that the method works
as intended even if the base entity page doesn't exist. A PageIdentity
doesn't have to be an existing page. This fixes an edge case where
::onLinksUpdateComplete() may erroneously try to save data to a /Votes
page because we originally checked for existence of the base page, and
returned the given $identity if it didn't exist. See comments at
I11b34a5578.

WishlistConfig::getCanonicalEntityPageRef() now handles Votes pages too.
Consequently, ::getVotesPageRefForEntity() no longer returns null when
given a votes page, which should be fine and seems logical.

Note that, critically, we return a Title object (which implements
PageIdentity) in getCanonicalEntityPage(), but we don't cast it to a new
PageIdentity as that would query for the page ID. Effectively, the page
ID is now lazy-loaded, keeping ::getCanonicalEntityPage() suitable for
calling in purely static contexts.

Change-Id: I289d9ddad61c5c7a9a541b3e93acca4b946eaa9c