WishlistTestTrait: always set page_lang after creating page
MediaWikiIntegrationTestCase::insertPage() constructs a WikiPage from
the title and loses the page language in the process. So, if we want to
do assertions involving the page_lang field, we need to manually set it
after the page is created.

See comments at I6ed2b01cba for why this is needed.

Reinforce with a new test that tests the tests! Wow!

Change-Id: Ib0f48983e5beff707a6d2dcc50da9c7df896ceeb