stores: distinguish "virtual" fields from translatable and wikitext
Add ::getMappedFields() to be more explicit about which fields need to
be mapped to column names in the storage layer, as opposed to
wikitext-only fields and translatable fields, both of which
*incidentally* are mapped to a column name.

Before getExtTranslateFields() was used for this purpose. In I391cd32e2d
the 'phabtasks' field was made into a wikitext-only field, and thus we
needed a virtual column. It was added to getExtTranslateFields() even
though it isn't actually translatable. This patch is basically a
follow-up to clear up that confusion.

In a future patch, we'll need to reference explicitly which fields may
contain <translate> tags, i.e. not 'phabtasks', hence why the new method
is added rather than renaming getExtTranslateFields().

Change-Id: Ie57989a67b48d515f8e2ac6e22621fcd3530342f