Ensure titles are sanitized for display and storage
TODO: Make the forms escape applicable wikitext like template syntax

We want to escape user input so that it appears and is stored just as it
was entered in the form. We have to make an exception for wrapping SPAN
tags that are added by Extension:Translate. Those contain 'lang' and
'dir' attributes that we want to preserve.

This involves some hackery in AbstractRenderer::getTitleSpan(). The lang
and dir attributes are parsed from any surrounding SPAN tag, then we
unwrap the SPAN and re-use the attributes on the SPAN we create.

Bug: T407961
Bug: T407000
Change-Id: I6e4fc43f3d1db9084375dea6e59f86dd3a5411a5