tests: turn AbstractWishlistEntityTest into MockWishlistConfigTrait
Traits are better for sharing logic in unit tests. This also matches the
existing paradigm of "MockWhateverTrait" that's seen across much of
MediaWiki unit tests.

This patch is in preparation of separating out hook handlers into
separate classes, which are all (mostly) tested in unit tests. This will
keep mocked configuration DRY and in one place.

Change-Id: I2caac67962bc3ab5a5ab0037f14903ed55298a4a