Escape wikitext syntax in wikitext storage, but keep it everywhere else
This is a follow-up to I6e4fc43f3d. Escape wikitext when saving to the
page, and unescape in AbstractRenderer so that it gets stored identical
to as it was entered. The HTML that gets set on the display title is
once again escaped by Html::element().

AbstractWishlistEntity::toWikitext() is the authoritative method for how
to generate wikitext from an entity, so we do the sanitization there.
This means it will get used when editing via the API as well.

The title is also unescaped when converting from wikitext, such as in
::toArray() and ::newFromWikitextParams().

In addition to 'dir' and 'lang', we should copy 'class' to our <span> as
the class is also needed by Extension:Translate.

Change tests to use {{gallery}} as the test template which uses the Core
tag <gallery> to better illustrate strip markers via template.
<references/> relies on Extension:Cite to be installed.

Bug: T407961
Bug: T407000
Follow-Up: I6e4fc43f3d1db9084375dea6e59f86dd3a5411a5
Depends-On: I8f8b1ecee2374cc81a119fe061b1ff4297eb7398
Change-Id: If6e803c17fd653365d96bb4d418e6e730fc5a9e4