Render everything possible in the user interface language
These changes effectively make everything in the user interface language
(such as with ?uselang) except user-submitted content. For those, we set
'lang' and 'dir' attributes so the MinT module can translate them. This
in effect drastically improves performance of the MinT integration, as
would be necessary for enabling it more broadly (T408160).

- Set the ParserOutput language in the ParserAfterTidy hook to the user
  interface language.
- Allow users of AbstractRenderer::getDivRaw() to specify a Language for
  setting the 'lang' and 'dir' attributes for user-submitted content.
- Render timestamps in the user interface language.
- Use AbstractRenderer::msg() instead of Parser::msg().
- Add a new prop to ApiQueryWishes, 'langinfo', which returns the lang
  and direction for the wish and its focus area (if applicable).
  This is then used to set those attributes in the wish-index table,
  removing the need for the getLangAttr() function.
- Fix TranslationBanner.vue to not set 'undefined' as the 'dir'
  attribute when switching back to original content

In a separate patch, similar tactics can be used by VoteRenderer so that
votes can easily be translated.

Bug: T407349
Change-Id: I98a35e72b7d0a28a2d84af2548ffbbf11a55b510