Invalidate cache when changing lang, and make base_lang rewritable
Invalidating the cache when changing page language should have been the
case all along. This should fix T409322, specifically, or at least
improve the situation.

Meanwhile, if a user submits a wish written in a language other than the
their interface language, a sysop has to manually correct the page
language and the base_lang parameter. However we weren't including
base_lang in the SET part of our UPDATE query, so in the database would
never get updated. This is now fixed, and similarly we include the
created_at timestamp for focus areas to be in the SET as you are
supposed to be able to change that.

This patch should fix these issues, but the larger issue is that users
can't indicate which language they are writing in when using the intake
form, and also that we wait until after the page is created to set the
language. That will be addressed in a separate patch.

Bug: T409992
Bug: T409322
Change-Id: I8df6e60199a5d10dd3a88e0654f046ea9dd93019