AbstractChangesProcessor: service for comparing two entities
This centralizes the logic for determining what changed between two
versions of an entity. AbstractChangesProcessor (suggestions for a
better name are most welcome) takes two entity objects and compiles the
differences into a structured format.

For now this service is only used for generating edit summaries, but in
subsequent patches will also be used for logging (T410810) and in edit
filters (T411051). The log entries will then be used for the "Updates"
feature and to send out notifications (T406286).

Bug: T410810
Bug: T411051
Bug: T406286
Change-Id: Idae5aedb094a907d34cf20003f10b82e60e2943a