Add logging for entity creation, status changes and FA changes
Adds a new entrypoint to EntityChangesProcessor that adds log entries
for the fields specified by ::getLogEntryFields(). Focus area titles are
added to ::getPreloadTitles() as applicable for LinkBatch optimization.

::addLogEntires is called by ApiWishlistEntityBase after a successful
save, meaning it will get the 'tag-community-wishlist' change tag
applied to it just like the edit.

In a future patch, we can send out notifications for logged events, and
transclude the log in the new "Updates" section (T406286).

Bug: T410810
Change-Id: I42f9e699183c56bb17684692c3f2298875cc602e