Notify page subscribers when notable changes are made
This leverages DiscussionTools to send out notifications to page
subscribers when:
* A wish status changes
* A wish's focus area changes
* A focus area status changes
* A wish is added or removed from a focus area

For now, we use the same design as DiscussionTools in the notification.
That is, it is presented as a 'notice' (not an 'alert'), using the
'edit' icon (since the notification was triggered by an edit), and with
simple messaging. The notifications are categorized under
'communityrequests' and should be grouped together if there are several.
A priority of 4 is given, which puts it as the same priority as a
notification where a user was pinged.

AbstractPresentationModel is the main model for all notifications, and
we have subclasses for each of the four types.

Bug: T411841
Change-Id: I346608017b02329eefb02fae9cd3c36ac553e148